+++ src/states/elasticache-replication-group.state.ts
@@ -1,0 +1,8 @@
+import AWS from 'aws-sdk';
+
+export type ElasticacheReplicationGroupState = {
+  id: string;
+  status: string;
+  snapshotName: string;
+  createParams: AWS.ElastiCache.CreateReplicationGroupMessage;
+};